requirements.txt
Django==1.6.2
Markdown==2.4
South==0.8.4
boto==2.27.0
dj-database-url==0.2.2
dj-static==0.0.5
django-storages==1.1.8
django-toolbelt==0.0.1
djangorestframework==2.3.13
drf-extensions==0.2.2
fake-factory==0.4.0
gunicorn==18.0
psycopg2==2.5.1
requests==2.3.0
static==0.4
wsgiref==0.1.2
